+++ app/javascript/components/fsp-maps/reducers.js
@@ -171,6 +171,28 @@ export default {
     const analysis = { ...state.analysis, nearby };
     return { ...state, analysis };
   },
+  [actions.toggleLocation]: (state, { payload }) => {
+    const nearby = { ...state.analysis.nearby, location: payload };
+    const analysis = { ...state.analysis, nearby };
+    return { ...state, analysis };
+  },
+
+  // Analysis - Analyse patterns
+  [actions.setLocationsList]: (state, { payload }) => {
+    const location = { ...state.analysis.location, list: payload };
+    const analysis = { ...state.analysis, location };
+    return { ...state, analysis };
+  },
+  [actions.setLocationSelected]: (state, { payload }) => {
+    const location = { ...state.analysis.location, selectedLocation: payload };
+    const analysis = { ...state.analysis, location };
+    return { ...state, analysis };
+  },
+  [actions.setLocationArea]: (state, { payload }) => {
+    const location = { ...state.analysis.location, area: payload };
+    const analysis = { ...state.analysis, location };
+    return { ...state, analysis };
+  },
 
   // Analysis - area of interest
   [actions.setAreaOfInterest]: (state, { payload }) => {